+++ src/app/server-cluster/tests/BUILD.gn
@@ -22,6 +22,7 @@ chip_test_suite("tests") {
     "TestAttributeListBuilder.cpp",
     "TestDefaultServerCluster.cpp",
     "TestServerClusterInterfaceRegistry.cpp",
+    "TestSingleEndpointServerClusterRegistry.cpp",
   ]
 
   cflags = [ "-Wconversion" ]
